[Feature] Adding REST methods that allow to use the app as an API for listing (unsure if inserts or delete works remotly)

+++ persistence/src/main/java/dao/ComputerDao.java
@@ -23,7 +23,7 @@ public interface ComputerDao {
    */
   List<Computer> getComputerFromName(Computer computer);
   
-  List<Computer> getComputer(int id);
+  List<Computer> getComputer(long id);
 
   List<Computer> orderBy(Page page);
